# frontend/api_doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Movie Review System API",
    "description": "Backend API for Movie Review Website",
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/folder/update": {
      "put": {
        "tags": [
          "收藏夹管理"
        ],
        "summary": "更新收藏夹",
        "description": "修改收藏夹的名称、描述或公开状态",
        "operationId": "updateFolder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dto": {
                    "$ref": "#/components/schemas/FavoriteFolderDTO"
                  },
                  "user": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/folder/create": {
      "post": {
        "tags": [
          "收藏夹管理"
        ],
        "summary": "创建收藏夹",
        "description": "创建一个新的自定义收藏夹",
        "operationId": "createFolder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dto": {
                    "$ref": "#/components/schemas/FavoriteFolderDTO"
                  },
                  "user": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultFavoriteFolder"
                }
              }
            }
          }
        }
      }
    },
    "/favorite/add": {
      "post": {
        "tags": [
          "收藏管理"
        ],
        "summary": "添加收藏",
        "description": "将电影加入用户的收藏夹",
        "operationId": "addFavorite",
        "parameters": [
          {
            "name": "movieId",
            "in": "query",
            "description": "电影ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "folderId",
            "in": "query",
            "description": "收藏夹ID，不传则加入默认收藏夹",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/folder/{folderId}": {
      "get": {
        "tags": [
          "收藏夹管理"
        ],
        "summary": "获取收藏夹详情",
        "description": "根据ID获取收藏夹的详细信息",
        "operationId": "getFolderById",
        "parameters": [
          {
            "name": "folderId",
            "in": "path",
            "description": "收藏夹ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultFavoriteFolder"
                }
              }
            }
          }
        }
      }
    },
    "/folder/my": {
      "get": {
        "tags": [
          "收藏夹管理"
        ],
        "summary": "获取我的收藏夹列表",
        "description": "获取当前用户的所有收藏夹",
        "operationId": "getMyFolders",
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListFavoriteFolderVO"
                }
              }
            }
          }
        }
      }
    },
    "/folder/count": {
      "get": {
        "tags": [
          "收藏夹管理"
        ],
        "summary": "统计收藏夹数量",
        "description": "获取用户的收藏夹总数",
        "operationId": "countFolders",
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultInteger"
                }
              }
            }
          }
        }
      }
    },
    "/favorite/status": {
      "get": {
        "tags": [
          "收藏管理"
        ],
        "summary": "查询收藏状态",
        "description": "查询当前用户是否收藏了某部电影",
        "operationId": "isFavorited",
        "parameters": [
          {
            "name": "movieId",
            "in": "query",
            "description": "电影ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/favorite/my": {
      "get": {
        "tags": [
          "收藏管理"
        ],
        "summary": "获取我的收藏列表",
        "description": "分页获取当前用户收藏的所有电影，包含收藏时间",
        "operationId": "getMyFavorites",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页数量",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageInfoMyFavoriteVO"
                }
              }
            }
          }
        }
      }
    },
    "/favorite/folder/{folderId}": {
      "get": {
        "tags": [
          "收藏管理"
        ],
        "summary": "获取收藏夹内电影",
        "description": "分页获取指定收藏夹内的电影列表",
        "operationId": "getFolderMovies",
        "parameters": [
          {
            "name": "folderId",
            "in": "path",
            "description": "收藏夹ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页数量",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageInfoMyFavoriteVO"
                }
              }
            }
          }
        }
      }
    },
    "/favorite/count": {
      "get": {
        "tags": [
          "收藏管理"
        ],
        "summary": "获取收藏总数",
        "description": "统计当前用户的收藏电影总数",
        "operationId": "getFavoritesCount",
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultInteger"
                }
              }
            }
          }
        }
      }
    },
...
    "/folder/delete/{folderId}": {
      "delete": {
        "tags": [
          "收藏夹管理"
        ],
        "summary": "删除收藏夹",
        "description": "删除收藏夹及其下的所有收藏记录",
        "operationId": "deleteFolder",
        "parameters": [
          {
            "name": "folderId",
            "in": "path",
            "description": "收藏夹ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/favorite/remove": {
      "delete": {
        "tags": [
          "收藏管理"
        ],
        "summary": "取消收藏",
        "description": "将电影从用户的收藏夹移除",
        "operationId": "removeFavorite",
        "parameters": [
          {
            "name": "movieId",
            "in": "query",
            "description": "电影ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "folderId",
            "in": "query",
            "description": "收藏夹ID，指定则只从该收藏夹移除",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/favorite/clear": {
      "delete": {
        "tags": [
          "收藏管理"
        ],
        "summary": "清空收藏夹",
        "description": "清空当前用户的所有收藏",
        "operationId": "clearMyFavorites",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/favorite/batch": {
      "delete": {
        "tags": [
          "收藏管理"
        ],
        "summary": "批量删除收藏",
        "description": "批量移除收藏夹中的多部电影",
        "operationId": "deleteFavoritesBatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "movieIds": {
                    "type": "array",
                    "description": "电影ID列表",
                    "items": {
                      "type": "integer",
                      "format": "int64"
                    }
                  },
                  "user": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户ID (唯一账号)",
            "example": "movie_fan_01"
          },
          "nickname": {
            "type": "string",
            "description": "用户昵称",
            "example": "电影狂热者"
          },
          "avatar": {
            "type": "string",
            "description": "用户头像 URL",
            "example": "http://localhost:8080/images/default_avatar.png"
          },
          "url": {
            "type": "string",
            "description": "个人主页/豆瓣链接",
            "example": "https://www.douban.com/people/xxxx"
          },
          "email": {
            "type": "string",
            "description": "电子邮箱",
            "example": "user@example.com"
          },
          "role": {
            "type": "integer",
            "description": "用户角色 (0:管理员, 1:普通用户)",
            "format": "int32",
            "example": 1
          },
          "status": {
            "type": "integer",
            "description": "账号状态 (0:正常, 1:被禁用/冻结)",
            "format": "int32",
            "example": 0
          },
          "passwordVersion": {
            "type": "integer",
            "description": "密码版本号，用于失效旧 Token（修改密码后递增）",
            "format": "int32",
            "example": 1
          },
          "createTime": {
            "type": "string",
            "description": "账号创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "最后更新时间",
            "format": "date-time"
          }
        },
        "description": "用户实体"
      },
      "ResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "type": "string",
            "description": "业务数据载荷"
          }
        },
        "description": "统一 API 响应结构"
      },
      "FavoriteFolderDTO": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "收藏夹ID（更新时需要）",
            "format": "int64",
            "example": 1
          },
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "description": "收藏夹名称",
            "example": "我的科幻电影"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "description": "收藏夹描述",
            "example": "收藏各种科幻题材的电影"
          },
          "isPublic": {
            "type": "integer",
            "description": "是否公开：0-私密, 1-公开",
            "format": "int32",
            "example": 0
          }
        },
        "description": "创建/更新收藏夹请求参数"
      },
...
      "FavoriteFolder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "收藏夹ID",
            "format": "int64",
            "example": 1
          },
          "userId": {
            "type": "string",
            "description": "用户ID",
            "example": "movie_fan_01"
          },
          "name": {
            "type": "string",
            "description": "收藏夹名称",
            "example": "我的科幻电影"
          },
          "description": {
            "type": "string",
            "description": "收藏夹描述",
            "example": "收藏各种科幻题材的电影"
          },
          "isPublic": {
            "type": "integer",
            "description": "是否公开：0-私密, 1-公开",
            "format": "int32",
            "example": 0
          },
          "movieCount": {
            "type": "integer",
            "description": "电影数量",
            "format": "int32",
            "example": 15
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间",
            "format": "date-time"
          }
        },
        "description": "用户自定义收藏夹"
      },
      "ResultFavoriteFolder": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/FavoriteFolder"
          }
        },
        "description": "统一 API 响应结构"
      },
      "ResultBoolean": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "type": "boolean",
            "description": "业务数据载荷"
          }
        },
        "description": "统一 API 响应结构"
      },
...
      "MyFavoriteVO": {
        "type": "object",
        "properties": {
          "movieId": {
            "type": "integer",
            "description": "电影ID",
            "format": "int64",
            "example": 1292052
          },
          "movieName": {
            "type": "string",
            "description": "电影名称",
            "example": "肖申克的救赎"
          },
          "cover": {
            "type": "string",
            "description": "电影封面URL",
            "example": "http://localhost:8080/images/p480747492.jpg"
          },
          "score": {
            "type": "number",
            "description": "豆瓣评分 (0-10)",
            "format": "double",
            "example": 9.7
          },
          "votes": {
            "type": "integer",
            "description": "评分人数",
            "format": "int32",
            "example": 2956885
          },
          "genres": {
            "type": "string",
            "description": "电影类型 (逗号分隔)",
            "example": "犯罪,剧情"
          },
          "year": {
            "type": "integer",
            "description": "上映年份",
            "format": "int32",
            "example": 1994
          },
          "directors": {
            "type": "array",
            "description": "导演列表",
            "example": [
              {
                "name": "弗兰克·德拉邦特",
                "id": "15615"
              }
            ],
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "object",
                "description": "导演列表",
                "example": [
                  {
                    "name": "弗兰克·德拉邦特",
                    "id": "15615"
                  }
                ]
              },
              "description": "导演列表",
              "example": [
                {
                  "name": "弗兰克·德拉邦特",
                  "id": "15615"
                }
              ]
            }
          },
          "actors": {
            "type": "array",
            "description": "演员列表",
            "example": [
              {
                "name": "蒂姆·罗宾斯",
                "id": "12345"
              }
            ],
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "object",
                "description": "演员列表",
                "example": [
                  {
                    "name": "蒂姆·罗宾斯",
                    "id": "12345"
                  }
                ]
              },
              "description": "演员列表",
              "example": [
                {
                  "name": "蒂姆·罗宾斯",
                  "id": "12345"
                }
              ]
            }
          },
          "regions": {
            "type": "string",
            "description": "制片国家/地区",
            "example": "美国"
          },
          "mins": {
            "type": "string",
            "description": "片长",
            "example": "142分钟"
          },
          "storyline": {
            "type": "string",
            "description": "剧情简介",
            "example": "20世纪40年代末，小有成就的青年银行家安迪..."
          },
          "favoriteTime": {
            "type": "string",
            "description": "收藏时间",
            "example": "2024-01-15 14:30:00"
          }
        },
        "description": "我的收藏记录 (包含电影信息和收藏时间)"
      },
      "PageInfoMyFavoriteVO": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MyFavoriteVO"
            }
          },
          "pageNum": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "startRow": {
            "type": "integer",
            "format": "int64"
          },
          "endRow": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int32"
          },
          "prePage": {
            "type": "integer",
            "format": "int32"
          },
          "nextPage": {
            "type": "integer",
            "format": "int32"
          },
          "isFirstPage": {
            "type": "boolean"
          },
          "isLastPage": {
            "type": "boolean"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "navigatePages": {
            "type": "integer",
            "format": "int32"
          },
          "navigatepageNums": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "navigateFirstPage": {
            "type": "integer",
            "format": "int32"
          },
          "navigateLastPage": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "业务数据载荷"
      },
      "ResultPageInfoMyFavoriteVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/PageInfoMyFavoriteVO"
          }
        },
        "description": "统一 API 响应结构"
      },
      "ResultInteger": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "type": "integer",
            "description": "业务数据载荷",
            "format": "int32"
          }
        },
        "description": "统一 API 响应结构"
      },
      "FavoriteFolderVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "收藏夹ID",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "收藏夹名称",
            "example": "我的科幻电影"
          },
          "description": {
            "type": "string",
            "description": "收藏夹描述",
            "example": "收藏各种科幻题材的电影"
          },
          "isPublic": {
            "type": "integer",
            "description": "是否公开：0-私密, 1-公开",
            "format": "int32",
            "example": 0
          },
          "movieCount": {
            "type": "integer",
            "description": "电影数量",
            "format": "int32",
            "example": 15
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "example": "2024-01-15 14:30:00"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间",
            "example": "2024-02-07 10:00:00"
          }
        },
        "description": "收藏夹详情视图对象"
      },
      "ResultListFavoriteFolderVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "type": "array",
            "description": "业务数据载荷",
            "items": {
              "$ref": "#/components/schemas/FavoriteFolderVO"
            }
          }
        },
        "description": "统一 API 响应结构"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}